requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
browser-cookie3==0.20.1
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
fonttools==4.59.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.9
lxml==6.0.0
lz4==4.4.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
numpy==2.2.6
olefile==0.47
openai==1.98.0
orjson==3.11.1
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pdfcrowd==6.5.2
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pyhtml2pdf==0.0.7
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
selenium==4.34.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.42
streamlit==1.47.1
svglib==1.5.1
tenacity==9.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
watchdog==6.0.0
weasyprint==66.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xhtml2pdf==0.2.17
yarl==1.20.1
zopfli==0.2.3.post1
zstandard==0.23.0
